requirements.txt
numpy<2
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
SQLAlchemy==2.0.36
psycopg[binary]
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pdfplumber==0.11.4
python-docx==1.1.2
sentence-transformers==3.0.1 
boto3
psycopg2-binary



